refactor: 서랍장 메인 화면 배너 이미지 교체 (#151)
* refactor: 기존 soomsil 로고 파일 삭제

* refactor: 서랍장 배너 이미지를 S3에 업로드된 이미지로 교체

* chore: `soomsil_v2_logo` -> `soomsil_logo`로 파일 이름 변경
